# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Battery Intelligence Platform API
  description: |
    Comprehensive REST API for US Battery Industry data, analytics, and forecasting.

    ## Features
    - Company and facility data with advanced filtering
    - Geospatial queries for facility locations
    - Market forecasts and cost projections
    - Technology analytics and trends
    - Policy impact analysis
    - Natural language query interface (RAG-powered chatbot)

    ## Authentication
    All endpoints require API key authentication via `X-API-Key` header.

    ## Rate Limiting
    - Free tier: 100 requests/hour
    - Standard tier: 1,000 requests/hour
    - Enterprise tier: 10,000 requests/hour

  version: 1.0.0
  contact:
    name: Battery Research Team
    email: api@batteryintel.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.batteryintel.com/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
...
security:
  - ApiKeyAuth: []
paths:
...
  /companies:
    get:
      tags:
        - Companies
      summary: List all companies
      description: Get a paginated list of battery companies with optional filtering
      parameters:
        - name: technology
          in: query
          description: Filter by battery technology (e.g., "Li-ion", "LFP", "Solid-state")
          schema:
            type: string
            example: "LFP"
        - name: state
          in: query
          description: Filter by state abbreviation
          schema:
            type: string
            example: "NV"
        - name: stage
          in: query
          description: Filter by development stage
          schema:
            type: string
            enum: [announced, construction, operational, expansion]
        - name: min_capacity
          in: query
          description: Minimum capacity in GWh
          schema:
            type: number
            minimum: 0
            example: 50
        - name: max_capacity
          in: query
          description: Maximum capacity in GWh
          schema:
            type: number
            example: 200
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort_by
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, capacity, state]
            default: capacity
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /companies/{id}:
    get:
      tags:
        - Companies
      summary: Get company details
      description: Retrieve detailed information about a specific company
      parameters:
        - name: id
          in: path
          required: true
          description: Company ID or slug
          schema:
            type: string
            example: "tesla"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /facilities:
    get:
      tags:
        - Facilities
      summary: List all facilities
      description: Get a paginated list of manufacturing facilities with optional filtering
      parameters:
        - name: state
          in: query
          description: Filter by state abbreviation
          schema:
            type: string
            example: "TN"
        - name: company
          in: query
          description: Filter by company name or ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by operational status
          schema:
            type: string
            enum: [announced, under_construction, operational, expansion, decommissioned]
        - name: min_capacity
          in: query
          description: Minimum capacity in GWh
          schema:
            type: number
            minimum: 0
        - name: technology
          in: query
          description: Filter by technology type
          schema:
            type: string
        - name: bbox
          in: query
          description: Bounding box for geospatial filtering (minLon,minLat,maxLon,maxLat)
          schema:
            type: string
            example: "-125.0,24.0,-66.0,49.0"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /facilities/{id}:
    get:
      tags:
        - Facilities
      summary: Get facility details
      description: Retrieve detailed information about a specific facility
      parameters:
        - name: id
          in: path
          required: true
          description: Facility ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: string
          example: "tesla"
        name:
          type: string
          example: "Tesla"
        capacity:
          type: number
          description: Total capacity in GWh
          example: 110
        technology:
          type: string
          example: "Li-ion (4680, 2170)"
        facilities:
          type: array
          items:
            type: string
        states:
          type: array
          items:
            type: string
          example: ["NV", "TX"]
        partnerships:
          type: array
          items:
            type: string
        website:
          type: string
          format: uri
    CompanyDetail:
      allOf:
        - $ref: '#/components/schemas/Company'
        - type: object
          properties:
            description:
              type: string
            founded:
              type: integer
            headquarters:
              type: object
              properties:
                city:
                  type: string
                state:
                  type: string
                country:
                  type: string
            facilities_detail:
              type: array
              items:
                $ref: '#/components/schemas/Facility'
            financial_data:
              type: object
              properties:
                market_cap:
                  type: number
                revenue:
                  type: number
                investment:
                  type: number
            key_executives:
              type: array
              items:
                type: object
    CompanyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Company'
        pagination:
          $ref: '#/components/schemas/Pagination'
        filters_applied:
          type: object
    Facility:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Tesla Gigafactory Nevada"
        company:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        capacity:
          type: number
          description: Capacity in GWh
          example: 100
        technology:
          type: string
        status:
          type: string
          enum: [announced, under_construction, operational, expansion, decommissioned]
        operational_date:
          type: string
          format: date
        investment:
          type: number
          description: Investment in USD
        employees:
          type: integer
    FacilityDetail:
      allOf:
        - $ref: '#/components/schemas/Facility'
        - type: object
          properties:
            description:
              type: string
            production_lines:
              type: integer
            energy_source:
              type: string
            certifications:
              type: array
              items:
                type: string
            supply_chain:
              type: object
    FacilityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Facility'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        country:
          type: string
          default: "USA"
        coordinates:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid query parameters"
            timestamp: "2025-11-09T12:00:00Z"
            path: "/api/v1/companies"
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or missing API key"
            timestamp: "2025-11-09T12:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            timestamp: "2025-11-09T12:00:00Z"
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per hour
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when the rate limit resets (Unix timestamp)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit exceeded. Try again later."
            timestamp: "2025-11-09T12:00:00Z"